+++ packages/manager/src/features/Images/ImagesCreate/CreateImageTab.tsx
@@ -18,13 +18,14 @@ import DiskSelect from 'src/features/linodes/DiskSelect';
 import LinodeSelect from 'src/features/linodes/LinodeSelect';
 import { useGrants, useProfile } from 'src/queries/profile';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
-// import calculateCostFromUnitPrice from 'src/utilities/calculateCostFromUnitPrice';
-// import { convertStorageUnit } from 'src/utilities/unitConversions';
+import calculateCostFromUnitPrice from 'src/utilities/calculateCostFromUnitPrice';
+import { convertStorageUnit } from 'src/utilities/unitConversions';
 import getAPIErrorFor from 'src/utilities/getAPIErrorFor';
 import ImagesPricingCopy from './ImagesPricingCopy';
 import withImages, {
   ImagesDispatch,
 } from 'src/containers/withImages.container';
+import useFlags from 'src/hooks/useFlags';
 
 const useStyles = makeStyles((theme: Theme) => ({
   helperText: {
@@ -162,15 +163,17 @@ export const CreateImageTab: React.FC<Props & ImagesDispatch> = (props) => {
   };
 
   const requirementsMet = checkRequirements();
-  // TODO: uncomment for billing on 9/1
-  /* const selectedDiskData: Disk | undefined = disks.find(
-   *   (d) => `${d.id}` === selectedDisk
-   * );
-   * const selectedDiskSizeInGB = convertStorageUnit(
-   *   'MB',
-   *   selectedDiskData?.size,
-   *   'GB'
-   * ); */
+
+  const flags = useFlags();
+  const isImagePricingEnabled = Boolean(flags.imagesPriceInfo);
+  const selectedDiskData: Disk | undefined = disks.find(
+    (d) => `${d.id}` === selectedDisk
+  );
+  const selectedDiskSizeInGB = convertStorageUnit(
+    'MB',
+    selectedDiskData?.size,
+    'GB'
+  );
 
   const hasErrorFor = getAPIErrorFor(
     {
@@ -231,6 +234,14 @@ export const CreateImageTab: React.FC<Props & ImagesDispatch> = (props) => {
           disabled={!canCreateImage}
           data-qa-disk-select
         />
+        {isImagePricingEnabled ? (
+          <Typography className={classes.helperText} variant="body1">
+            {`Estimated: ${calculateCostFromUnitPrice(
+              0.1,
+              selectedDiskSizeInGB
+            )}/month`}
+          </Typography>
+        ) : null}
         <Typography className={classes.helperText} variant="body1">
           Linode Images cannot be created if you are using raw disks or disks
           that have been formatted using custom filesystems.